+++ gradio/layouts/__init__.py
@@ -6,6 +6,7 @@ from .group import Group
 from .row import Row
 from .sidebar import Sidebar
 from .tabs import Tab, TabItem, Tabs
+from .walkthrough import Step, Walkthrough
 
 __all__ = [
     "Accordion",
@@ -18,4 +19,6 @@ __all__ = [
     "Tab",
     "TabItem",
     "Sidebar",
+    "Walkthrough",
+    "Step",
 ]