+++ js/app/src/routes/[...catchall]/+page.svelte
@@ -239,7 +239,7 @@
 	let pending_deep_link_error = false;
 
 	let gradio_dev_mode = "";
-	let i18n_ready: boolean;
+
 	onMount(async () => {
 		//@ts-ignore
 		config = data.config;
@@ -385,6 +385,7 @@
 	pages={config.pages}
 	current_page={config.current_page}
 	root={config.root}
+	components={config.components}
 	loaded={loader_status === "complete"}
 	fill_width={config?.fill_width || false}
 	bind:wrapper