+++ js/component-test/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Dependency updates
 
+- @self/build@0.3.0
+- @gradio/video@0.15.1
+
+## 0.5.31
+
+### Dependency updates
+
 - @gradio/video@0.15.0
 
 ## 0.5.30